Moved withConfirm to misc, It could enhance any component that has an onclick handler, not only toolbar buttons
